// sourcecode/BTrak.Packages/snova-training-matrix/projects/project-components/src/lib/training-matrix/dependencies/models/training-course-model.ts
export class TrainingCourseModel {
    id: string;
    courseName: string;
    courseDescription: string;
    validityInMonths: number;
    companyId: string;
    isArchived: boolean;
    createdDateTime: Date;
    updatedDateTime: Date;
    createdByUserId: string;
    // isArchived: boolean;
    // expenseId: string;
    // description: string;
    // expenseCategoryId: string;
    // paymentStatusId: string;
    // cashPaidThroughId: string;
    // expenseReportId: string;
    // expenseStatusId: string;
    // billReceiptId: string;
    // claimReimbursement: boolean;
    // merchantId: string;
    // receiptDate: Date;
    // expenseDate: Date;
    // amount: number;
    // repliedByUserId: string;
    // repliedDateTime: Date;
    // reason: string;
    // isApproved: boolean;
    // isPaid: boolean;
    // isRecurringExpense: boolean = false;
    // actualBudget: number;
    // referenceNumber: string;
    // currencyId: string;
    // branchId: string;
    // timeStamp: any;
    // sortBy: any;
    // sortDirectionAsc: any;
    // searchText: string;
    // pageSize: number;
    // pageNumber: number;
    // to: string[];
    // cc: string[];
    // bcc: string[];
    // cronExpression: string;
    // cronExpressionId: string;
    // cronExpressionTimeStamp: any;
    // jobId: string;
    // claimedByUserId: string;
    // expenseName:string;
}